requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.0.0
google-genai>=1.0.0
pyserial>=3.5
